requirements.txt
discord.py
jmespath
reportlab
